+++ tests/bunit.core.tests/TestServiceProviderTest.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections;
 using System.Linq;
+using System.Threading.Tasks;
 using Microsoft.AspNetCore.Components;
 using Microsoft.Extensions.DependencyInjection;
 using Shouldly;
@@ -208,6 +209,52 @@ namespace Bunit
 
 			// Act and assert
 			Should.NotThrow(() => ctx.RenderComponent<DummyComponentWhichRequiresDummyService>());
+		}
+
+		[Fact(DisplayName = "Can correctly resolve and dispose of scoped disposable service")]
+		public void Test031()
+		{
+			var sut = new TestServiceProvider();
+			sut.AddScoped<DisposableService>();
+			var disposable = sut.GetService<DisposableService>();
+
+			sut.Dispose();
+
+			disposable.IsDisposed.ShouldBeTrue();
+		}
+
+		[Fact(DisplayName = "Can correctly resolve and dispose of transient disposable service")]
+		public void Test032()
+		{
+			var sut = new TestServiceProvider();
+			sut.AddTransient<DisposableService>();
+			var disposable = sut.GetService<DisposableService>();
+
+			sut.Dispose();
+
+			disposable.IsDisposed.ShouldBeTrue();
+		}
+
+		[Fact(DisplayName = "Can correctly resolve and dispose of singleton disposable service")]
+		public void Test033()
+		{
+			var sut = new TestServiceProvider();
+			sut.AddSingleton<DisposableService>();
+			var disposable = sut.GetService<DisposableService>();
+
+			sut.Dispose();
+
+			disposable.IsDisposed.ShouldBeTrue();
+		}
+
+		private sealed class DisposableService : IDisposable
+		{
+			public bool IsDisposed { get; private set; }
+
+			public void Dispose()
+			{
+				IsDisposed = true;
+			}
 		}
 	}
 }