+++ tests/bunit.core.tests/TestServiceProviderTest.net5.cs
@@ -1,6 +1,5 @@
-#if !NETCOREAPP3_1
+#if NET5_0_OR_GREATER
 using System;
-using System.Diagnostics.CodeAnalysis;
 using System.Threading.Tasks;
 using Microsoft.Extensions.DependencyInjection;
 using Shouldly;
@@ -10,20 +9,51 @@ namespace Bunit
 {
 	public partial class TestServiceProviderTest
 	{
-		[Fact(DisplayName = "Can correctly dispose of async disposable service")]
-		[SuppressMessage("Reliability", "CA2000:Dispose objects before losing scope", Justification = "Point of test is to verify explicit call to Dispose doesn't throw.")]
+		[Fact(DisplayName = "Can correctly resolve and dispose of scoped disposable service")]
 		public void Net5Test001()
 		{
 			var sut = new TestServiceProvider();
 			sut.AddScoped<AsyncDisposableService>();
-			sut.GetService<AsyncDisposableService>();
+			var asyncDisposable = sut.GetService<AsyncDisposableService>();
 
-			Should.NotThrow(() => sut.Dispose());
+			sut.Dispose();
+
+			asyncDisposable.IsDisposed.ShouldBeTrue();
 		}
 
-		private class AsyncDisposableService : IAsyncDisposable
+		[Fact(DisplayName = "Can correctly resolve and dispose of transient disposable service")]
+		public void Net5Test002()
 		{
-			public ValueTask DisposeAsync() => ValueTask.CompletedTask;
+			var sut = new TestServiceProvider();
+			sut.AddTransient<AsyncDisposableService>();
+			var asyncDisposable = sut.GetService<AsyncDisposableService>();
+
+			sut.Dispose();
+
+			asyncDisposable.IsDisposed.ShouldBeTrue();
+		}
+
+		[Fact(DisplayName = "Can correctly resolve and dispose of singleton disposable service")]
+		public void Net5Test003()
+		{
+			var sut = new TestServiceProvider();
+			sut.AddSingleton<AsyncDisposableService>();
+			var asyncDisposable = sut.GetService<AsyncDisposableService>();
+
+			sut.Dispose();
+
+			asyncDisposable.IsDisposed.ShouldBeTrue();
+		}
+
+		private sealed class AsyncDisposableService : IAsyncDisposable
+		{
+			public bool IsDisposed { get; private set; }
+
+			public ValueTask DisposeAsync()
+			{
+				IsDisposed = true;
+				return ValueTask.CompletedTask;
+			}
 		}
 	}
 }